Remove unused init and fix imports
